+++ confignode/src/main/java/org/apache/iotdb/confignode/conf/ConfigNodeConfCheck.java
@@ -18,7 +18,8 @@
  */
 package org.apache.iotdb.confignode.conf;
 
-import org.apache.iotdb.confignode.exception.conf.RepeatConfigurationException;
+import org.apache.iotdb.commons.exception.ConfigurationException;
+import org.apache.iotdb.commons.exception.StartupException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -48,7 +49,7 @@ public class ConfigNodeConfCheck {
     specialProperties = new Properties();
   }
 
-  public void checkConfig() throws RepeatConfigurationException, IOException {
+  public void checkConfig() throws ConfigurationException, IOException, StartupException {
     File specialPropertiesFile =
         new File(conf.getSystemDir() + File.separator + ConfigNodeConstant.SPECIAL_CONF_NAME);
     if (!specialPropertiesFile.exists()) {
@@ -62,7 +63,7 @@ public class ConfigNodeConfCheck {
         LOGGER.error(
             "Can't create special configuration file {} for ConfigNode. IoTDB-ConfigNode is shutdown.",
             specialPropertiesFile.getAbsolutePath());
-        throw new IOException("Can't create special configuration file");
+        throw new StartupException("Can't create special configuration file");
       }
     }
 
@@ -89,13 +90,13 @@ public class ConfigNodeConfCheck {
   }
 
   /** Ensure that special parameters are consistent with each startup except the first one */
-  private void checkSpecialProperties() throws RepeatConfigurationException {
+  private void checkSpecialProperties() throws ConfigurationException {
     int specialDeviceGroupCount =
         Integer.parseInt(
             specialProperties.getProperty(
                 "device_group_count", String.valueOf(conf.getDeviceGroupCount())));
     if (specialDeviceGroupCount != conf.getDeviceGroupCount()) {
-      throw new RepeatConfigurationException(
+      throw new ConfigurationException(
           "device_group_count",
           String.valueOf(conf.getDeviceGroupCount()),
           String.valueOf(specialDeviceGroupCount));
@@ -106,7 +107,7 @@ public class ConfigNodeConfCheck {
             "device_group_hash_executor_class", conf.getDeviceGroupHashExecutorClass());
     if (!Objects.equals(
         specialDeviceGroupHashExecutorClass, conf.getDeviceGroupHashExecutorClass())) {
-      throw new RepeatConfigurationException(
+      throw new ConfigurationException(
           "device_group_hash_executor_class",
           conf.getDeviceGroupHashExecutorClass(),
           specialDeviceGroupHashExecutorClass);
