+++ integration/src/test/java/org/apache/iotdb/db/integration/IoTDBCheckConfigIT.java
@@ -18,10 +18,10 @@
  */
 package org.apache.iotdb.db.integration;
 
+import org.apache.iotdb.commons.exception.ConfigurationException;
 import org.apache.iotdb.db.conf.IoTDBConfigCheck;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.engine.fileSystem.SystemFileFactory;
-import org.apache.iotdb.db.exception.ConfigurationException;
 import org.apache.iotdb.db.service.IoTDB;
 import org.apache.iotdb.db.utils.EnvironmentUtils;
 import org.apache.iotdb.itbase.category.LocalStandaloneTest;
